better organization through abstraction
