+++ controllers/helpers.js
@@ -1,0 +1,19 @@
+// HELPER FUNCTIONS:
+
+// Hash converter for sqlite look ups
+exports.convertHash = hash => {
+  let x = parseInt(hash);
+  if (x > 0xFFFFFFFF) {
+    console.error('Too big, must have a wrong number');
+  }
+  if (x > 0x7FFFFFFF) {
+    x = 0x100000000 - x;
+    if (x < 2147483648) {
+      return -x
+    }
+    else {
+      return -2147483648
+    }
+  }
+  return x;
+}
