+++ controllers/itemHandlers.js
@@ -1,0 +1,3 @@
+exports.handleWeapons = async () => {
+  console.log('weaponHandler');
+}
